+++ demo/dam_break_2d/dam_break_2d_neighbour_list.jl
@@ -5,13 +5,6 @@
   @ description:
  =#
 
-#=
-  @ author: bcynuaa <bcynuaa@163.com> | callm1101 <Calm.Liu@outlook.com> | vox-1 <xur2006@163.com>
-  @ date: 2024/06/12 22:08:32
-  @ license: MIT
-  @ description:
- =#
-
 using EtherSPH
 using Parameters
 using ProgressBars